requirements.txt
pre-commit
pandas==2.2.3
bs4==0.0.2
lxml==5.3.1
python-dotenv==1.0.1
pydantic==2.10.6
langchain-community==0.3.18
paddleocr==2.10.0
paddlepaddle
tabulate==0.9.0
azure-ai-formrecognizer
azure.ai.documentintelligence==1.0.0
py-asciimath==0.3.0
transformers==4.49.0
streamlit==1.43.2